ocr/detection/ctpn/requirements.txt
numpy
scipy
pillow
easydict
opencv-python
keras
h5py
PyYAML
cython
tensorflow
myelin